requirements.txt
Flask==2.3.2
PyJWT==2.7.0
python-dotenv==1.0.0
